Project-2/Revisi/requirements.txt
pandas
kafka-python
pyspark
flask
flask-cors
scikit-learn
joblib
requests
streamlit